Add dockerfile build for the WeatherFinder app
I moved all the app sources from WeatherFinder to WeatherFinder/app.
I added a Makefile and a Dockerfile. These two files can be easily reused for the other apps.
I added a docker_entrypoint.sh script which is direcly executed from the container. I this case, it only launchs the akka application. Please keep it as simple as possible and add the compilation steps in the Dockerfile itself (see the RUN mvn compile line)
Please refer to README.md for more informations on how to create and push the docker image.